+++ packages/date-picker/src/DatePicker/DatePicker.tsx
@@ -1,5 +1,5 @@
 import React, { RefObject, useEffect, useRef, useState } from "react"
-import { format, parse } from "date-fns"
+import { parse } from "date-fns"
 import { DateRange, DateInterval, DayClickEventHandler } from "react-day-picker"
 import { FocusOn } from "react-focus-on"
 import { usePopper } from "react-popper"
@@ -10,6 +10,9 @@ import { isInvalidDate } from "../utils/isInvalidDate"
 import { isDisabledDate } from "../utils/isDisabledDate"
 import { setFocusInCalendar } from "../utils/setFocusInCalendar"
 import { formatDateAsText } from "../utils/formatDateAsText"
+import { formatDateAsNumeral } from "../utils/formatDateAsNumeral"
+import { getLocale } from "../utils/getLocale"
+import { SupportedLocales } from "../types"
 import calendarStyles from "./components/Calendar/Calendar.scss"
 import { DateFormat, DayOfWeek } from "./enums"
 import { Calendar, CalendarElement, CalendarProps } from "./components/Calendar"
@@ -25,6 +28,7 @@ type OmittedDateInputProps =
   | "onButtonClick"
   | "calendarId"
   | "value"
+  | "locale"
 
 export interface DatePickerProps
   extends Omit<DateInputProps, OmittedDateInputProps> {
@@ -35,6 +39,7 @@ export interface DatePickerProps
   onInputChange?: DateInputProps["onChange"]
   onInputBlur?: DateInputProps["onBlur"]
   onButtonClick?: DateInputProps["onButtonClick"]
+  locale: SupportedLocales
   /**
    * Accepts a DayOfWeek value to start the week on that day. By default,
    * it's set to Monday.
@@ -122,9 +127,11 @@ export type ValidationResponse = {
  * {@link https://cultureamp.design/components/date-picker/ Guidance} |
  * {@link https://cultureamp.design/storybook/?path=/docs/components-date-picker-date-picker--default-story Storybook}
  */
+
 export const DatePicker: React.VFC<DatePickerProps> = ({
   id,
   buttonRef = useRef<HTMLButtonElement>(null),
+  locale: propsLocale,
   disabledDates,
   disabledDaysOfWeek,
   disabledRange,
@@ -152,6 +159,8 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
     null
   )
 
+  const locale = getLocale(propsLocale)
+
   const [lastTrigger, setLastTrigger] = useState<
     "inputFocus" | "inputKeydown" | "calendarButton"
   >()
@@ -166,6 +175,7 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
       },
     ],
     placement: "bottom-start",
+    strategy: "fixed",
   })
 
   const disabledDays = calculateDisabledDays({
@@ -229,9 +239,9 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
   const handleCalendarDayChange: DayClickEventHandler = date => {
     if (!isDisabledDate(date, disabledDays)) {
       if (lastTrigger === "calendarButton") {
-        setInputValue(format(date, DateFormat.Text))
+        setInputValue(formatDateAsText(date, disabledDays, locale))
       } else {
-        setInputValue(format(date, DateFormat.Numeral))
+        setInputValue(formatDateAsNumeral(date, locale))
       }
 
       handleDayChange(date)
@@ -246,7 +256,12 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
 
   const handleInputFocus: React.FocusEventHandler<HTMLInputElement> = e => {
     setLastTrigger("inputFocus")
-    selectedDay && setInputValue(format(selectedDay, DateFormat.Numeral))
+    if (selectedDay) {
+      const newInputValue = isInvalidDate(selectedDay)
+        ? ""
+        : formatDateAsNumeral(selectedDay, locale)
+      setInputValue(newInputValue)
+    }
     onInputFocus && onInputFocus(e)
   }
 
@@ -262,10 +277,12 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
     if (isSelectingDayInCalendar) return
 
     if (inputValue !== "") {
-      const parsedDate = parse(inputValue, DateFormat.Numeral, new Date())
+      const parsedDate = parse(inputValue, DateFormat.Numeral, new Date(), {
+        locale,
+      })
 
       if (!isInvalidDate(parsedDate)) {
-        formatDateAsText(parsedDate, disabledDays, setInputValue)
+        setInputValue(formatDateAsText(parsedDate, disabledDays, locale))
       }
 
       handleDayChange(parsedDate, inputValue)
@@ -292,12 +309,9 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
   }
 
   const handleCalendarMount = (calendarElement: CalendarElement): void => {
-    switch (lastTrigger) {
-      case "inputFocus":
-        return
-      default:
-        setFocusInCalendar(calendarElement, selectedDay)
-    }
+    if (lastTrigger === "inputFocus") return
+
+    setFocusInCalendar(calendarElement, selectedDay)
   }
 
   const handleReturnFocus = (): void => {
@@ -309,7 +323,8 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
   }
 
   useEffect(() => {
-    selectedDay && formatDateAsText(selectedDay, disabledDays, setInputValue)
+    selectedDay &&
+      setInputValue(formatDateAsText(selectedDay, disabledDays, locale))
 
     if (selectedDay && isInvalidDate(selectedDay)) {
       onValidate({
@@ -326,7 +341,8 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
     }
 
     if (selectedDay && isDisabledDate(selectedDay, disabledDays)) {
-      const formattedDate = format(selectedDay, DateFormat.Numeral)
+      const formattedDate = formatDateAsNumeral(selectedDay, locale)
+
       onValidate({
         date: undefined,
         inputValue: formattedDate,
@@ -355,31 +371,33 @@ export const DatePicker: React.VFC<DatePickerProps> = ({
             buttonRef={buttonRef}
             id={id}
             calendarId={`${id}-calendar-dialog`}
+            value={inputValue}
+            locale={locale}
             isCalendarOpen={isOpen}
+            icon={dateStart}
+            onButtonClick={handleButtonClick}
             onClick={handleInputClick}
             onFocus={handleInputFocus}
             onChange={handleInputChange}
             onBlur={handleInputBlur}
             onKeyDown={handleKeyDown}
-            onButtonClick={handleButtonClick}
-            icon={dateStart}
-            value={inputValue}
             {...restDateInputProps}
           />
         </div>
         {isOpen && (
           <Calendar
-            mode="single"
             id={`${id}-calendar-dialog`}
-            setPopperElement={setPopperElement}
-            popperStyles={styles}
-            popperAttributes={attributes}
+            mode="single"
             value={selectedDay}
             defaultMonth={defaultMonth}
             weekStartsOn={weekStartsOn}
             disabledDays={disabledDays}
+            locale={locale}
+            popperStyles={styles}
+            popperAttributes={attributes}
             onDayChange={handleCalendarDayChange}
             onMount={handleCalendarMount}
+            setPopperElement={setPopperElement}
           />
         )}
       </div>