+++ packages/date-picker/src/DatePicker/components/DateInput/DateInput.tsx
@@ -9,7 +9,6 @@ import {
   InputProps,
   Label,
 } from "@kaizen/draft-form"
-import { Matcher } from "react-day-picker/src/types/Matchers"
 import { getDescription } from "./utils/getDescription"
 import styles from "./DateInput.scss"
 
@@ -25,6 +24,7 @@ type OmittedInputProps =
   | "ariaDescribedBy"
   | "defaultInputValue"
   | "automationId"
+  | "locale"
 
 export interface DateInputProps extends Omit<InputProps, OmittedInputProps> {
   buttonRef?: React.RefObject<HTMLButtonElement>
@@ -50,6 +50,7 @@ export interface DateInputProps extends Omit<InputProps, OmittedInputProps> {
    * A descriptive message for `status` states
    */
   validationMessage?: string | React.ReactNode
+  locale: Locale
 }
 
 export const DateInput: React.VFC<DateInputProps> = ({
@@ -67,9 +68,9 @@ export const DateInput: React.VFC<DateInputProps> = ({
   status,
   validationMessage,
   value,
+  locale,
   ...inputProps
 }) => {
-  // Focus behaviour breaks when this is a function component.
   const IconButton: React.ReactNode = (
     <button
       ref={buttonRef}
@@ -128,7 +129,7 @@ export const DateInput: React.VFC<DateInputProps> = ({
       >
         <FieldMessage
           id={descriptionId}
-          message={getDescription(description)}
+          message={getDescription(description, locale)}
           reversed={isReversed}
         />
       </div>